requirements.txt
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.0
board_to_fen==0.1.1
cachetools==6.2.2
certifi==2025.11.12
charset-normalizer==3.4.4
chess==1.11.2
click==8.3.1
fastapi==0.124.0
google-auth==2.43.0
google-genai==1.54.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.11
pillow==12.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.2
pydantic==2.12.5
pydantic_core==2.41.5
python-chess==1.999
requests==2.32.5
rsa==4.9.1
starlette==0.50.0
tenacity==9.1.2
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.6.1
uvicorn==0.38.0
websockets==15.0.1
